requirements.txt
langchain_community
langchain_chroma
langchainhub